// src/utils/types/dadosQuorum.types.ts
export interface DadosQuorum {
    cpf: string;
    foto: string;
    nome: string;
    percentualAlunosResp: number;
    percentualFuncionarios: number;
    percentualTotal: number;
    quantFuncionarios: number;
    quantRespAlunosNaoVotantes: number;
    quantRespAlunosVotantes: number;
    quantidadeAlunosVotantes: number;
    totalVotosAlunosResponsaveis: number;
    totalVotosAlunosResponsaveisFuncionarios: number;
    votosAlunos: number;
    votosFuncionarios: number;
    votosResponsaveisNaoVotantes: number;
    votosResponsaveisVotantes: number;

    // totalQtdAlunosVotantes: number;
    // percentualTotalAlunosVotantes: number;
    // totalQtdAlunosResponsaveisAlunosNaoVotantes: number;
    // precentualTotalQtdAlunosResponsaveisAlunosNaoVotantes: number;
    // totalQtdAlunosResponsaveisVotantes: number;
    // percentualtotalQtdAlunosResponsaveisVotantes: number;
    // totalQtdFuncionarios: number;
    // percentualTotalQtdFuncionarios: number;



}